+++ Drop-In/src/main/java/com/braintreepayments/api/VaultManagerFragment.java
@@ -1,6 +1,5 @@
 package com.braintreepayments.api;
 
-import android.content.DialogInterface;
 import android.os.Bundle;
 import android.os.Parcelable;
 import android.view.LayoutInflater;
@@ -9,7 +8,6 @@ import android.view.ViewGroup;
 import android.widget.ViewSwitcher;
 
 import androidx.annotation.VisibleForTesting;
-import androidx.appcompat.app.AlertDialog;
 import androidx.fragment.app.Fragment;
 import androidx.lifecycle.Observer;
 import androidx.lifecycle.ViewModelProvider;
@@ -19,7 +17,6 @@ import androidx.recyclerview.widget.RecyclerView;
 import com.braintreepayments.api.dropin.R;
 
 import java.util.List;
-import java.util.concurrent.atomic.AtomicBoolean;
 
 import static com.braintreepayments.api.DropInUIEventType.DISMISS_VAULT_MANAGER;
 
@@ -75,46 +72,9 @@ public class VaultManagerFragment extends Fragment implements View.OnClickListen
         if (v instanceof PaymentMethodItemView) {
             PaymentMethodItemView paymentMethodItemView = (PaymentMethodItemView) v;
             final PaymentMethodNonce paymentMethodNonceToDelete = paymentMethodItemView.getPaymentMethodNonce();
-            PaymentMethodItemView dialogView = new PaymentMethodItemView(requireActivity());
-            dialogView.setPaymentMethod(paymentMethodNonceToDelete, false);
 
-            final AtomicBoolean positiveSelected = new AtomicBoolean(false);
-            new AlertDialog.Builder(requireActivity(),
-                    R.style.Theme_AppCompat_Light_Dialog_Alert)
-                    .setTitle(R.string.bt_delete_confirmation_title)
-                    .setMessage(R.string.bt_delete_confirmation_description)
-                    .setView(dialogView)
-                    .setPositiveButton(R.string.bt_delete, new DialogInterface.OnClickListener() {
-                        @Override
-                        public void onClick(DialogInterface dialog, int which) {
-                            positiveSelected.set(true);
-
-                            DeleteVaultedPaymentMethodNonceEvent event = new DeleteVaultedPaymentMethodNonceEvent(paymentMethodNonceToDelete);
-                            sendBraintreeEvent(event);
-                            // TODO: this is required because there is a delay on the back end after
-                            //  deleting a payment method, so when immediately refetching nonces it
-                            //  still exists. Updating the adapter directly removes the deleted nonce
-                            //  from the view. This is also why the alert dialog lives in the fragment
-                            //  instead of the activity. Is there a better way to do this?
-                            adapter.paymentMethodDeleted(paymentMethodNonceToDelete);
-                        }
-                    })
-                    .setOnDismissListener(new DialogInterface.OnDismissListener() {
-                        @Override
-                        public void onDismiss(DialogInterface dialog) {
-                            if (!positiveSelected.get()) {
-                                // TODO: Is this analytics event important? It requires another event to be dispatched and handled in DropInActivity
-//                                getDropInClient().sendAnalyticsEvent("manager.delete.confirmation.negative");
-                            }
-                        }
-                    })
-                    .setNegativeButton(R.string.bt_cancel, new DialogInterface.OnClickListener() {
-                        @Override
-                        public void onClick(DialogInterface dialog, int which) {
-                        }
-                    })
-                    .create()
-                    .show();
+            DeleteVaultedPaymentMethodNonceEvent event = new DeleteVaultedPaymentMethodNonceEvent(paymentMethodNonceToDelete);
+            sendBraintreeEvent(event);
         }
     }
